+++ srcs/Server.cpp
@@ -16,9 +16,9 @@ void	Server::set_server_sock(/*config info*/)
 {
 	//slightly less temp, how do we get the right server object from the socket in question?
 	const std::vector<ConfigServer*>& servers = m_Config.m_servers;
-	for(int i = 0; i < servers.size(); i++){
-		const std::vector<std::pair<std::string,std::string>>& ports = m_Config.m_servers[i]->getPorts();
-		for(int j = 0;j < ports.size(); j++){
+	for(size_t i = 0; i < servers.size(); i++){
+		const std::vector<std::pair<std::string,std::string> >& ports = m_Config.m_servers[i]->getPorts();
+		for(size_t j = 0;j < ports.size(); j++){
 			Socket sock = Socket(AF_INET, SOCK_STREAM, 0);
 			PRINT_LOG("server [",i,"] Binding socket at:",ports[j].second );
 			sock.bind_socket(ports[j].first, std::atoi(ports[j].second.c_str()));
@@ -31,7 +31,7 @@ void	Server::set_server_sock(/*config info*/)
 
 void	Server::start_listening()//need to add poll later
 {
-	for (int n = 0; n < m_serverSockVec.size(); n++)
+	for (size_t n = 0; n < m_serverSockVec.size(); n++)
 	{
 		m_serverSockVec[n].enable_listener();
 	}
@@ -40,7 +40,7 @@ void	Server::start_listening()//need to add poll later
 
 void	Server::accept_connection()
 {
-	for (int n = 0; n < m_serverSockVec.size(); n++)
+	for (size_t n = 0; n < m_serverSockVec.size(); n++)
 	{
 		pollfd	sfd = {m_serverSockVec[n].get_sock_fd(), POLL_IN, 0};
 		m_pfdVec.push_back(sfd);
@@ -56,7 +56,7 @@ void	Server::accept_connection()
 			exit(EXIT_FAILURE);
 		}
 		// std::cout << "result = " << result << std::endl;
-		for (int n = 0; n < m_pfdVec.size(); n++)
+		for (size_t n = 0; n < m_pfdVec.size(); n++)
 		{
 			if ((m_pfdVec[n].revents & POLLIN) && !isClientSock(m_pfdVec[n].fd))
 			{
@@ -113,7 +113,7 @@ void Server::handle( Socket& client_sock )
 
 bool	Server::isClientSock(int fdToCheck)
 {
-	for (int  n = 0; n < m_serverSockVec.size(); n++)
+	for (size_t  n = 0; n < m_serverSockVec.size(); n++)
 	{
 		if (m_serverSockVec[n].get_sock_fd() == fdToCheck)
 			return (false);
@@ -148,7 +148,7 @@ void	Server::removeFromVec(int fdToRemove)
 
 std::string	Server::getHostFromFd(int fdToFind)
 {
-	for (int n = 0; n < m_serverSockVec.size(); n++)
+	for (size_t n = 0; n < m_serverSockVec.size(); n++)
 	{
 		if (m_serverSockVec[n].get_sock_fd() == fdToFind)
 		{
@@ -156,7 +156,7 @@ std::string	Server::getHostFromFd(int fdToFind)
 		}
 	}
 
-	for (int n = 0; n < m_clientSockVec.size(); n++)
+	for (size_t n = 0; n < m_clientSockVec.size(); n++)
 	{
 		if (m_clientSockVec[n].get_sock_fd() == fdToFind)
 		{
@@ -169,7 +169,7 @@ std::string	Server::getHostFromFd(int fdToFind)
 
 Socket&	Server::getServerSockFromVec(int fdToFind)
 {
-	for (int n = 0; n < m_serverSockVec.size(); n++)
+	for (size_t n = 0; n < m_serverSockVec.size(); n++)
 	{
 		if (m_serverSockVec[n].get_sock_fd() == fdToFind)
 			return (m_serverSockVec[n]);
@@ -179,7 +179,7 @@ Socket&	Server::getServerSockFromVec(int fdToFind)
 
 Socket&	Server::getClientSockFromVec(int fdToFind)
 {
-	for (int n = 0; n < m_clientSockVec.size(); n++)
+	for (size_t n = 0; n < m_clientSockVec.size(); n++)
 	{
 		if (m_clientSockVec[n].get_sock_fd() == fdToFind)
 			return (m_clientSockVec[n]);
